Add words.xml to GIT
Formely words.xml has had to be in the exe-dir which has been ruled out
from the GIT repo by .gitignore ignoring the bin dir. Now words.xml is
searched from any parent directory of the exe-dir, allowing it to be set
to the main repo dir and added by git.